UI: Move tunnel issue generation code to TunnelIssue.js
